Move more update logic to client side
Similar to the previous move of `status`, return a structure
from the daemon and render it on the client.  This actually
moves more logic for updating to the client too; the IPC request
updates just one component, and the client iterates over all
installed components to update.

This way we also better return "partial" status.