cli: add flag to control logging level
This introduces a new `-v(vv)` flag on both `bootupd` and `bootupctl`
to control log verbosity. The default level only shows warnings and
higher priorities.
The service unit sets up the daemon to log info messages too, and
can be overridden via a drop-in.
This also introduces some basic unit-tests on CLI parsing.