daemon: Exit when done processing requests
There's no reason to sit there taking up RAM just because
someone did `bootupctl status`.  Properly implementing
exit-on-idle really requires an event loop (e.g. `tokio`)
but that'd be a large change.

For now let's just exit when a client disconnects.  However,
this quickly broke the test suite which does a lot of
commands in a row.

Tweak the systemd unit to allow more restarts, and also
add a few strategic sleeps.

Closes: https://github.com/coreos/bootupd/issues/43