packaging: add a systemd unit to run every boot
This enables bootloader updates automatically on boot.
Note that the service is intentionally not enabled by default,
it should be up to the distribution to add a systemd preset if
auto-update for the bootloader is desired.

Right now RAID setups are not supported but see [1] for an example
in coreos.

[1] https://github.com/coreos/fedora-coreos-config/pull/3042