main: Don't use package system to query our own version
When we were reworking `bootc install` to try to run outside
of a container, we hit on the fact that this query required
rpm in the *host* context which is conceptually broken.

As part of severing our dependency on rpm (and package systems)
in general let's just use our own crate version which we know
at compile time.

Signed-off-by: Colin Walters <walters@verbum.org>