tests/e2e-update: drop manifest_fork logic
I feel like this was a bit heavyweight for what we were gaining
from it since we were only modifying one file.

It just so happens also that the `git clone` was not including the
git submodule that we're now using in FCOS for bootc.

Let's just drop the "forking" and do something a bit more lightweight.