component: Make update metadata generation to be optional
Currently, the `bootupctl backend generate-update-metadata` command tries
to generate an update layout for each component, and it bails out if one
of the components can't be installed.

There are systems in which some of the bootupd components are not needed
(e.g.: x86_64 EFI-only systems). But if one of the requirements for one
of the components is not meet (e.g.: grub2-install is missing for BIOS),
this will lead to a fatal error instead of just skipping the component
that is not supported.

To prevent this, let's make the update layout generation to be optional.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>